Check if the browser supports Flash
One function included: "_fnHasFlash"

returns a boolean:
- true if the browser supports Flash
- false if not

Then, I edited one statement of "_fnCreateButton" function. Target: not to create any Flash button if the browser does not support it.

Please, check it, it is very useful for mobile devices which does not support Flash... This way, every Flash button will be shown as text button.

Thanks to albertein for publishing this way to check Flash capabilities. See StackOverflow post: http://stackoverflow.com/questions/998245/how-can-i-detect-if-flash-is-installed-and-if-not-display-a-hidden-div-that-inf